+++ src/sql/exceptions.py
@@ -47,7 +47,7 @@ NotImplementedError = exception_factory("NotImplementedError")
 # raised internally when the user chooses a table that doesn't exist
 TableNotFoundError = exception_factory("TableNotFoundError")
 
-# raise it when there is an error in parsing pyproject.toml file
+# raise it when there is an error in parsing the configuration file
 ConfigurationError = exception_factory("ConfigurationError")