+++ src/tests/test_config.py
@@ -91,7 +91,44 @@ drivername = sqlite
     assert ConnectionManager.current.dialect == "sqlite"
 
 
-def test_start_ini_default_connection_using_pyproject_if_any(tmp_empty, ip_no_magics):
+def test_load_home_toml_if_no_pyproject_toml(
+    tmp_empty, ip_no_magics, capsys, monkeypatch
+):
+    monkeypatch.setattr(
+        Path, "expanduser", lambda path: Path(str(path).replace("~", tmp_empty))
+    )
+    home_toml = Path("~/.jupysql/config").expanduser()
+    home_toml.parent.mkdir(exist_ok=True)
+    home_toml.write_text(
+        """
+[tool.jupysql.SqlMagic]
+autocommit = false
+autolimit = 1
+style = "RANDOM"
+"""
+    )
+
+    expect = [
+        "Settings changed:",
+        r"autocommit\s*\|\s*False",
+        r"autolimit\s*\|\s*1",
+        r"style\s*\|\s*RANDOM",
+    ]
+
+    config_expected = {"autocommit": False, "autolimit": 1, "style": "RANDOM"}
+
+    os.mkdir("sub")
+    os.chdir("sub")
+
+    load_ipython_extension(ip_no_magics)
+    magic = ip_no_magics.find_magic("sql").__self__
+    combined = {**get_default_testing_configs(magic), **config_expected}
+    out, _ = capsys.readouterr()
+    assert all(re.search(substring, out) for substring in expect)
+    assert get_current_configs(magic) == combined
+
+
+def test_start_ini_default_connection_using_toml_if_any(tmp_empty, ip_no_magics):
     Path("pyproject.toml").write_text(
         """
 [tool.jupysql.SqlMagic]
@@ -130,12 +167,35 @@ drivername = someunknowndriver
     assert "Cannot start default connection" in captured.out
 
 
-def test_magic_initialization_with_no_pyproject(tmp_empty, ip_no_magics):
+def test_magic_initialization_with_no_toml(tmp_empty, ip_no_magics):
     load_ipython_extension(ip_no_magics)
 
 
-def test_magic_initialization_with_corrupted_pyproject(tmp_empty, ip_no_magics, capsys):
+def test_magic_initialization_with_corrupted_pyproject_toml(
+    tmp_empty, ip_no_magics, capsys
+):
     Path("pyproject.toml").write_text(
+        """
+[tool.jupysql.SqlMagic]
+dsn_filename = myconnections.ini
+"""
+    )
+
+    load_ipython_extension(ip_no_magics)
+
+    captured = capsys.readouterr()
+    assert "Could not load configuration file" in captured.out
+
+
+def test_magic_initialization_with_corrupted_home_toml(
+    tmp_empty, ip_no_magics, capsys, monkeypatch
+):
+    monkeypatch.setattr(
+        Path, "expanduser", lambda path: Path(str(path).replace("~", tmp_empty))
+    )
+    home_toml = Path("~/.jupysql/config").expanduser()
+    home_toml.parent.mkdir(exist_ok=True)
+    home_toml.write_text(
         """
 [tool.jupysql.SqlMagic]
 dsn_filename = myconnections.ini
@@ -184,6 +244,45 @@ style = "RANDOM"
     assert get_current_configs(magic) == combined
 
 
+def test_loading_valid_home_toml_shows_feedback_and_modifies_config(
+    tmp_empty, ip_no_magics, capsys, monkeypatch
+):
+    monkeypatch.setattr(
+        Path, "expanduser", lambda path: Path(str(path).replace("~", tmp_empty))
+    )
+    home_toml = Path("~/.jupysql/config").expanduser()
+    home_toml.parent.mkdir(exist_ok=True)
+    home_toml.write_text(
+        """
+[tool.jupysql.SqlMagic]
+autocommit = false
+autolimit = 1
+style = "RANDOM"
+"""
+    )
+
+    expect = [
+        "Loading configurations from {path}",
+        "Settings changed:",
+        r"autocommit\s*\|\s*False",
+        r"autolimit\s*\|\s*1",
+        r"style\s*\|\s*RANDOM",
+    ]
+
+    config_expected = {"autocommit": False, "autolimit": 1, "style": "RANDOM"}
+
+    os.mkdir("sub")
+    os.chdir("sub")
+
+    load_ipython_extension(ip_no_magics)
+    magic = ip_no_magics.find_magic("sql").__self__
+    combined = {**get_default_testing_configs(magic), **config_expected}
+    out, _ = capsys.readouterr()
+    expect[0] = expect[0].format(path=re.escape(str(home_toml)))
+    assert all(re.search(substring, out) for substring in expect)
+    assert get_current_configs(magic) == combined
+
+
 @pytest.mark.parametrize(
     "file_content, param",
     [
@@ -197,7 +296,7 @@ style = "RANDOM"
     ],
     ids=["empty_sqlmagic_key", "missing_sqlmagic_key"],
 )
-def test_loading_pyproject_toml_display_configuration_docs_link(
+def test_loading_toml_display_configuration_docs_link(
     tmp_empty, ip_no_magics, file_content, param, monkeypatch
 ):
     Path("pyproject.toml").write_text(file_content)
@@ -239,7 +338,7 @@ github = "ploomber/jupysql"
         ),
     ],
 )
-def test_load_pyproject_toml_user_configurations_not_specified(
+def test_load_toml_user_configurations_not_specified(
     tmp_empty, ip_no_magics, capsys, file_content
 ):
     Path("pyproject.toml").write_text(file_content)
@@ -359,6 +458,6 @@ autocommit = true
     out, _ = capsys.readouterr()
     assert (
         "The 'toml' package isn't installed. "
-        "To load settings from the pyproject.toml file, "
+        "To load settings from pyproject.toml or ~/.jupysql/config, "
         "install with: pip install toml"
     ) in out