+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.10.3dev
 
+* [Feature] Allow user-level config using ~/.jupysql/config (#880)
 * [Fix] Remove force deleted snippets from dependent snippet's `with` (#717)
 * [Fix] Comments added in SQL query to be stripped before saved as snippet (#886)
 * [Fix] Fixed bug passing :NUMBER while string slicing in query (#901)